Feat : Lyrics 추가, SongServiceImpl 수정

+++ Backend/jayul/src/main/java/com/example/song/repository/MainsongRepository.java
@@ -3,6 +3,6 @@ package com.example.song.repository;
 import com.example.song.domain.Mainsong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MainsongRepository extends JpaRepository<Mainsong, Integer> {
+public interface MainsongRepository extends JpaRepository<Mainsong, Long> {
 
 }
